Move nav_stack widget to gui service (was in busy app)
