+++ applications/services/gui/modules/nav_stack.c
